Resume Game/Prefilled Value Bugfix
Previously, the board determined whether a value was prefilled based on if that cell stored a value when drawing the cell. Now, that determination is done off of the initial state of the board, so when a user resumes a game with values filled, those values can still be changed.